permissions

+++ app/Http/Controllers/Admin/PlanController.php
@@ -41,7 +41,9 @@ class PlanController extends Controller
 
     public function show($url)
     {
-        $plan = $this->repository->where('url',$url)->first();
+        $plan = $this->repository
+        ->with('details')
+        ->where('url',$url)->first();
         if(!$plan)
             return redirect()->back();
 
@@ -54,13 +56,18 @@ class PlanController extends Controller
     public function destroy($url)
     {
         $plan = $this->repository
+                        ->with('details')
                         ->where('url', $url)
                         ->first();
 
         if (!$plan)
             return redirect()->back();
 
-
+        if ($plan->details->count() > 0) {
+            return redirect()
+                        ->back()
+                        ->with('error', 'Existem detahes vinculados a esse plano, portanto não pode deletar');
+        }
 
         $plan->delete();
 
@@ -90,7 +97,7 @@ class PlanController extends Controller
             'plan' => $plan
         ]);
     }
-    public function store(Request $request)
+    public function store(StoreUpdatePlan $request)
     {
         $data = $request->all();
         $data['url'] =Str::kebab($request->name) ;
